+++ app/Modules/Utilities/Controllers/BackupController.php
@@ -12,6 +12,10 @@
 namespace BT\Modules\Utilities\Controllers;
 
 use BT\Http\Controllers\Controller;
+use BT\Modules\Clients\Models\Client;
+use BT\Modules\Expenses\Models\Expense;
+use BT\Modules\RecurringInvoices\Models\RecurringInvoice;
+use BT\Modules\TimeTracking\Models\TimeTrackingProject;
 use BT\Support\DateFormatter;
 use Ifsnop\Mysqldump\Mysqldump;
 use Illuminate\Http\Request;
@@ -94,6 +98,40 @@ class BackupController extends Controller
         ini_set('max_execution_time', $maxtime);
 
         return back()->with('alertSuccess', trans('bt.record_successfully_deleted'));
+    }
+
+    public function clientInactivePrior (Request $request) {
+        $maxtime = ini_get('max_execution_time');
+        ini_set('max_execution_time', 0);
+
+        $date = DateFormatter::unformat($request->clientprior_date);
+        // find all clients with activity after $date
+        $active_quote = Quote::distinct('client_id')->where('quote_date', '>=', $date)->pluck('client_id','client_id');
+        $active_workorder = Workorder::distinct('client_id')->where('workorder_date', '>=', $date)->pluck('client_id','client_id');
+        $active_invoice = Invoice::distinct('client_id')->where('invoice_date', '>=', $date)->pluck('client_id','client_id');
+        $active_recurringinvoice = RecurringInvoice::distinct('client_id')->where('next_date', '>=', $date)->pluck('client_id','client_id');
+        $active_payment = Payment::distinct('client_id')->where('paid_at', '>=', $date)->pluck('client_id','client_id');
+        $active_expense = Expense::distinct('client_id')->where('expense_date', '>=', $date)->pluck('client_id','client_id');
+        $active_project = TimeTrackingProject::distinct('client_id')->where('due_at', '>=', $date)->pluck('client_id','client_id');
+        //union above
+        $results = $active_quote->union($active_workorder)->union($active_invoice)->union($active_recurringinvoice)
+            ->union($active_payment)->union($active_expense)->union($active_project);
+        //get all clients
+        $clients = Client::pluck('id','id');
+        //get client ids to set as inactive
+        $inactiveclients = $clients->diff($results);
+
+        $setinactive = Client::whereIn('id', $inactiveclients)->where('active', 1)->get();
+
+        foreach ($setinactive as $client){
+            $client->active = 0;
+            $client->save();
+        }
+
+        ini_set('max_execution_time', $maxtime);
+
+        return back()->with('alertSuccess', $setinactive->count().' Clients set to Inactive');
+
     }
 
 }