+++ app/Modules/Documentation/Views/documentation/Release-Notes.md
@@ -10,6 +10,10 @@ Release Notes
 - replaced bootstrap-switch with bootstrap-switch-button
 - added Schedule page for scheduled employees and resources
 - fix Vendor delete
+- added category, itemlookup and mail queue datatables
+- updated all datatables
+- add swalConfirm target (for sweetalert over iframe)
+- added bulk client inactive to database management
 
 ### BillingTrack 5.1.0 (2019-10-01)
 - Added client shipping address fields (address_2,city_2,etc) accessible in templates